change of format for "extra_volumes"
`extra_volumes` used to be string containing [] and was never used.
This was a mistake as it do not really make any sense from a json point of view.

For the backup feature `extra_volumes` will be used, we changed it in order to
contains useful information. It is now a dict.

All your existing images will now display `"extra_volumes": {}`, this is an api
break but as the previous value was useless we hope it will not disturb you too
much.