Amend volumes in server API
If `volume_type` is not spelled correctly, it results in an obscure API error:
```{
    "fields": {
        "volumes.1.size": [
            "extra keys not allowed"
        ],
        "volumes.1.organization": [
            "extra keys not allowed"
        ],
        "volumes.1.volume_type": [
            "extra keys not allowed"
        ],
        "volumes.1.id": [
            "required key not provided"
        ]
    },
    "message": "Validation Error",
    "type": "invalid_request_error"
}```
The key `volumes` is actually optional and will either create a new volume or use an existing volume depending on how the key is used. It is required for X64 servers that have a minimum allowance of 200GB of storage to be created but cannot use the 200GB on the root volume.